backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
supabase==2.4.6
python-dotenv==1.0.1
httpx==0.27.0
transformers>=4.40.0
torch
sentencepiece
accelerate
google-genai
